requirements.txt
streamlit
speechrecognition
transformers
torch
requests
beautifulsoup4
faiss-cpu
openai-whisper
sounddevice
scipy
